+++ CHANGELOG.md
@@ -11,8 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add sonatype publishing
 
 ### Changed
-- Kotlin to 1.4.30 stable
-- Added intermediate jsCommon main/test sourcesSet for node plugin
+- Kotlin to 1.4.30 stable.
+- Added intermediate jsCommon main/test sourcesSet for node plugin.
+- Plugin names changed to `ru.mipt.npm` package.
+- Common plugin id changed to `common`
+- Plugins group changed to `ru.mipt.npm` with `gradle prefix
 
 ### Deprecated